Fix save/load waypoint controls behaving inconsistently
When save button was last pushed, then any cartridge put into the port
afterwards was automatically ejected. This does no longer happen.
They are now only ejected if save/load button was pushed or the
port was punched.
Cartridges will also no longer be written to the moment they are put
into the port. They will only be written to if save button was pushed.
That means empty cartridges can be ejected empty.