fix telepathic circuits finding structures for free in Nether and the End
When the mallet was not used yet, then searching for structures in the Nether 
or the End via the telepathic circuits is for free, while it is not when 
searching for them in the Overworld.
This commit fixes that by multiplying the base cost of the Nether and End search 
by instability instead (which is always at least 1).
This brings it in line with the Overworld structure search.